Fix reconcile of deleted resources in namespaced mode
The predicate function for Delete events of the lockedresources reconciler assumes the client is a cluster scoped client with permissions to Get Namespace resources. If the call to get a Namespace fails, the deleted resources will not be reconciled. This commit fixes that by trying to reconcile the deleted resource even if the call to get the Namespace fails. This causes no problems as the k8s API will deny any create/update operation in a Namespace that's marked for termination (which was the original intention of the failing code)